move: Positive VO를 extractor -> common 이동
