+++ src/main/java/calculator/calculatorApplication/domain/calculation/NumberParser.java
@@ -1,6 +1,6 @@
 package calculator.calculatorApplication.domain.calculation;
 
-import calculator.calculatorApplication.domain.parser.vo.PositiveNumbers;
+import calculator.calculatorApplication.domain.common.PositiveNumbers;
 import java.util.List;
 
 public interface NumberParser {
