+++ src/main/java/calculator/calculatorApplication/domain/calculation/StringCalculator.java
@@ -1,7 +1,7 @@
 package calculator.calculatorApplication.domain.calculation;
 
 import calculator.calculatorApplication.Calculator;
-import calculator.calculatorApplication.domain.parser.vo.PositiveNumbers;
+import calculator.calculatorApplication.domain.common.PositiveNumbers;
 import java.util.List;
 
 public class StringCalculator implements Calculator {
