+++ src/test/java/calculator/caulculatorApplication/domain/parser/NumberParserTest.java
@@ -1,5 +1,6 @@
 package calculator.caulculatorApplication.domain.parser;
 
+import calculator.caulculatorApplication.domain.common.PositiveNumbersTest;
 import java.util.List;
 
 public interface NumberParserTest {
